Rename ButtonWidget folder to BottunWidgets and with this rename the namespace to reflect the content
